automatically backed up by learn

+++ index.js
@@ -5,4 +5,5 @@ function divide(a,b) {return a / b}
  add (8,6);
 function inc(n) {return n + 1}
 function dec(n) {return n - 1}
-function makeInt(n) {return parseInt(n,10)}
+function makeInt(n) {return parseFloat(n,10)}
+
